--- src/main/java/g2/tss/fyre/controllers/CreateController.java
@@ -1,13 +1,0 @@
-package g2.tss.fyre.controllers;
-
-/**
- * This interface describe creation of interface
- *
- * @author Anton Kudryavtsev
- */
-public interface CreateController {
-    /**
-     * This method create controller
-     */
-    void create();
-}
