--- src/main/java/g2/tss/fyre/models/datastorage/DataStorage.java
@@ -1,16 +1,0 @@
-package g2.tss.fyre.models.datastorage;
-
-import g2.tss.fyre.models.entity.Authorization;
-
-/**
- * This interface describe data storage worker
- * @author Anton Kudryavtsev
- */
-public interface DataStorage {
-    /**
-     * Get authorization info by login
-     * @param login login
-     * @return authorization info
-     */
-    Authorization getAuthorization(String login);
-}
