+++ src/main/java/tss/g2/fyre/controllers/CreateController.java
@@ -1,0 +1,12 @@
+package tss.g2.fyre.controllers;
+
+/**
+ * This interface describe creation of interface.
+ *
+ * @author Anton Kudryavtsev
+ */
+public interface CreateController {
+
+  /** This method create controller. */
+  void create();
+}
