+++ src/main/java/tss/g2/fyre/models/actions/Action.java
@@ -1,0 +1,11 @@
+package tss.g2.fyre.models.actions;
+
+import tss.g2.fyre.models.Answer;
+
+/**
+ * M.
+ * @author Anton Kudryavtsev
+ */
+public interface Action {
+  Answer getAnswer();
+}
